backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
bcrypt==4.1.3
python-dotenv==1.0.1
